move downloader to separate file
